requirements.txt
groq
edge-tts
pdfplumber
pydub
gradio
python-dotenv

 